+++ CHANGELOG.md
@@ -378,6 +378,7 @@
 * Fixed 2D collision matrix not matching 3D collision matrix
 * Added PathFollower entity that pushes actors along a path of NavigationNodes
 * Added ThingPusher entity that pushes other things along a path of NavigationNodes
+* Updated ForwardedUtils in AddonSupport and added a few missing methods
 * Added conditional/microscript resolvers for PlayerFlags and SessionFlags (conditional only) [Experimental]
 * Added ScriptStringSubber that calls scripts and substitutes the result [Experimental]
 * DelayScaled can now be used from threads other than the main thread [Untested]